change getPreferredSize()
git-svn-id: https://svn.code.sf.net/p/genj/code/trunk@1424 fbd980a2-0dff-11df-b4b7-0b05b74d0985

+++ genj/src/core/genj/util/swing/HeadlessLabel.java
@@ -128,11 +128,13 @@ public class HeadlessLabel extends JComponent {
    * @see javax.swing.JComponent#getPreferredSize()
    */
   public Dimension getPreferredSize() {
-    float width, height;
+    int 
+      width, 
+      height;
     // ask rootView or text
     if (view!=null) {
-      width = view.getPreferredSpan(View.X_AXIS);
-      height= view.getPreferredSpan(View.Y_AXIS);
+      width = (int)view.getPreferredSpan(View.X_AXIS);
+      height= (int)view.getPreferredSpan(View.Y_AXIS);
     } else {
       FontMetrics fm = getFontMetrics(getFont());
       width = fm.stringWidth(txt);
@@ -146,8 +148,12 @@ public class HeadlessLabel extends JComponent {
     // gap?
     if ((view!=null||txt.length()>0)&&icon!=null) 
       width += ICON_TEXT_GAP;
+    // max
+//    Dimension max = getMaximumSize();
+//    width = Math.min(max.width, width);
+//    height = Math.min(max.height, height);
     // done
-    return new Dimension((int)width,(int)height);
+    return new Dimension(width, height);
   }
     
   /**
